deck editor errors, styling. Delete card ID from deck broken

+++ frontend/src/app/services/file.service.ts
@@ -25,8 +25,8 @@ export class FileService {
     return this.webService.getWithArgs(`image/${fileName}`, {responseType: 'text'}, true);
   }
 
-  public remove(id: string): any {
-    return this.webService.delete(`file/del?id=${id}`, true);
+  public remove(id: string, deckName: string, userID: string): any {
+    return this.webService.delete(`file/del?id=${id}&deckName=${deckName}&userID=${userID}`, true);
     //console.log('file/del/'+fileName);
     // this.fileList.splice(this.fileList.findIndex(name => name === fileName), 1);
     // this.fileList$.next(this.fileList);
